+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonMessageBox.cs
@@ -338,6 +338,315 @@ public static class KryptonMessageBox
             showCtrlCopy: showCtrlCopy,
             showHelpButton: displayHelpButton,
             showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, string.Empty,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons, KryptonMessageBoxIcon icon,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons, icon,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons, icon, defaultButton,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons, icon, defaultButton, options,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <param name="helpInfo">Contains the help data of the <see cref="KryptonMessageBox"/>.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options, HelpInfo helpInfo,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons, icon, defaultButton, options, helpInfo,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="displayHelpButton">Displays a 'Help' button, as seen in .NET 6 and higher.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(string text, string caption, KryptonMessageBoxButtons buttons, bool displayHelpButton,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(null, text, caption, buttons, icon, defaultButton, options,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: displayHelpButton,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string text,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, string.Empty,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons, KryptonMessageBoxIcon icon,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons, icon,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons, icon, defaultButton,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons, icon, defaultButton, options,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: false,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <param name="helpInfo">Contains the help data of the <see cref="KryptonMessageBox"/>.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options, HelpInfo helpInfo,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons, icon, defaultButton, options,
+            showCtrlCopy: showCtrlCopy,
+            helpInfo: helpInfo,
+            showCloseButton: showCloseButton);
+
+    /// <summary>
+    /// Displays a message box asynchronously in front+center of the application and with the specified text, caption and buttons.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption" default="string.Empty">The text to display in the title bar of the message box. default="string.Empty"</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="displayHelpButton">Displays a 'Help' button, as seen in .NET 6 and higher.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    public static Task<DialogResult> ShowAsync(IWin32Window owner, string? text, string? caption, KryptonMessageBoxButtons buttons, bool displayHelpButton,
+        KryptonMessageBoxIcon icon, KryptonMessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+        bool? showCtrlCopy = null,
+        bool? showCloseButton = null) =>
+        ShowCoreAsync(owner, text, caption, buttons, icon, defaultButton, options,
+            showCtrlCopy: showCtrlCopy,
+            showHelpButton: displayHelpButton,
+            showCloseButton: showCloseButton);
+
     #endregion
 
     #region Implementation
@@ -390,6 +699,62 @@ public static class KryptonMessageBox
             kmb.StartPosition = showOwner == null ? FormStartPosition.CenterScreen : FormStartPosition.CenterParent;
 
             return kmb.ShowDialog(showOwner);
+        }
+    }
+
+    /// <summary>
+    /// Displays a message box asynchronously with the specified text, caption, buttons, icon, default button, options, and Help button, using the specified Help file, HelpNavigator, and Help topic.
+    /// </summary>
+    /// <param name="owner">Owner of the modal dialog box.</param>
+    /// <param name="text">The text to display in the message box.</param>
+    /// <param name="caption">The text to display in the title bar of the message box.</param>
+    /// <param name="buttons">One of the System.Windows.Forms.KryptonMessageBoxButtons values that specifies which buttons to display in the message box.</param>
+    /// <param name="icon">One of the KryptonMessageBoxIcon values that specifies which icon to display in the message box.</param>
+    /// <param name="defaultButton">One of the KryptonMessageBoxDefaultButton values that specifies the default button for the message box.</param>
+    /// <param name="options">One of the System.Windows.Forms.MessageBoxOptions values that specifies which display and association options will be used for the message box. You may pass in 0 if you wish to use the defaults.</param>
+    /// <param name="helpInfo">Contains the help data of the <see cref="KryptonMessageBox"/>.</param>
+    /// <param name="showCtrlCopy">Show extraText in title. If null (default) then only when Warning or Error icon is used.</param>
+    /// <param name="showHelpButton">Displays a 'Help' button, as seen in .NET 6 and higher.</param>
+    /// <param name="showCloseButton">Displays the close button. If null (default), then the close button will be displayed.</param>
+    /// <returns>A task that yields one of the System.Windows.Forms.DialogResult values.</returns>
+    private static async Task<DialogResult> ShowCoreAsync(IWin32Window? owner,
+        string? text, string? caption,
+        KryptonMessageBoxButtons buttons = KryptonMessageBoxButtons.OK,
+        KryptonMessageBoxIcon icon = KryptonMessageBoxIcon.None,
+        KryptonMessageBoxDefaultButton defaultButton = KryptonMessageBoxDefaultButton.Button1,
+        MessageBoxOptions options = 0,
+        HelpInfo? helpInfo = null,
+        bool? showCtrlCopy = null,
+        bool? showHelpButton = null,
+        bool? showCloseButton = null)
+    {
+        caption = string.IsNullOrEmpty(caption) ? @" " : caption;
+
+        IWin32Window? showOwner = ValidateOptions(owner, options, helpInfo);
+
+        // Show message box window as a modal dialog and then dispose of it after-wards
+
+        if (options is MessageBoxOptions.RightAlign or MessageBoxOptions.RtlReading)
+        {
+            using var kmbRtl = new VisualMessageBoxRtlAwareForm(showOwner, text, caption, buttons, icon,
+                defaultButton, helpInfo, showCtrlCopy, showHelpButton, showCloseButton);
+
+            kmbRtl.StartPosition = showOwner == null ? FormStartPosition.CenterScreen : FormStartPosition.CenterParent;
+
+            return showOwner is null
+                ? await kmbRtl.ShowDialogAsync().ConfigureAwait(true)
+                : await kmbRtl.ShowDialogAsync(showOwner).ConfigureAwait(true);
+        }
+        else
+        {
+            using var kmb = new VisualMessageBoxForm(showOwner, text, caption, buttons, icon,
+                defaultButton, helpInfo, showCtrlCopy, showHelpButton, showCloseButton);
+
+            kmb.StartPosition = showOwner == null ? FormStartPosition.CenterScreen : FormStartPosition.CenterParent;
+
+            return showOwner is null
+                ? await kmb.ShowDialogAsync().ConfigureAwait(true)
+                : await kmb.ShowDialogAsync(showOwner).ConfigureAwait(true);
         }
     }